+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## v2.10.1 - ?
+## v2.11.0 - ?
 
+- Automatically add JsonResource serialized entities to JsonFile resources when the serialized entity extends JsonResource
+- Allow to skip in-model serialization of JsonResource.
 
 ## v2.10.0 - 2026-06-20